feat(#5081): Refactor parsing logic and introduce new Parse class
